mock-service/requirements.txt
fastapi
uvicorn
pyyaml
requests
kafka-python
prometheus-fastapi-instrumentator
prometheus-client